+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Development
 
+## [0.1.6](https://github.com/berlinonline/ckanext-datasetsnippets/releases/tag/0.1.6)
+
+_(2024-05-29)_
+
 - The root element of the breadcrumb in the returned snippets can now be configured, either via a `root_breadcrumb`-URL parameter, or via the newly introduced config `datasetsnippets.default_root_breadcrumb`.
 - When setting config values in `plugin.py`, don't overwrite values that might have been set elsewhere.
 - Add help text for tags in facet list.